requirements.txt
redis~=7.1.1
aiogram~=3.24.0
python-dotenv==1.0.0
SQLAlchemy~=2.0.46
aiosqlite~=0.20.0
aio_pika
asyncpg